requirements.txt
streamlit==0.87.0
pandas==1.3.2
numpy==1.19.5
tensorflow==2.6.0
Pillow==8.3.1
